Manage NotFound exercise in use case instead of in MemoryRepo

+++ src/use-cases/exercise/DeleteExercise/DeleteExercise.usecase.ts
@@ -1,3 +1,4 @@
+import { NotFoundError } from '@/domain/common/errors';
 import { ExercisesRepo } from '@/domain/repos/ExercisesRepo.port';
 
 export type DeleteExerciseUsecaseRequest = {
@@ -8,6 +9,13 @@ export class DeleteExerciseUsecase {
   constructor(private exercisesRepo: ExercisesRepo) {}
 
   async execute(request: DeleteExerciseUsecaseRequest): Promise<void> {
+    // Search exercise
+    const exercise = await this.exercisesRepo.getExerciseById(request.id);
+
+    if (!exercise) {
+      throw new NotFoundError('DeleteExerciseUsecase: Exercise not found');
+    }
+
     await this.exercisesRepo.deleteExercise(request.id);
   }
 }
